fix mutex re-entrant locks. various bug fixes. input object for temperature profile. separate thread for temperature reading and controller update.

+++ src/common/temperature_profile.rs
@@ -1,5 +1,5 @@
-use serde::{Serialize, Deserialize};
-#[derive(PartialEq, Eq, Clone, Serialize, Deserialize)]
+use serde::Serialize;
+#[derive(PartialEq, Eq, Clone, Serialize)]
 pub enum ProfileState {
     Inactive,
     Pending,
@@ -7,15 +7,15 @@ pub enum ProfileState {
     Terminated
 }
 
-#[derive(Clone, Serialize, Deserialize)]
+#[derive(Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct TemperatureLevel {
     pub name: String,
     pub setpoint_c: f32,
     pub tolerance_c: f32,
-    pub duration_s: f32,
-    pub timer_s: f32,
-    pub start_time_s: f32,
+    pub duration_s: u64,
+    pub timer_s: u64,
+    pub start_time_s: u64,
     pub state: ProfileState,
 }
 
@@ -24,12 +24,12 @@ impl TemperatureLevel {
 
     pub fn init(&mut self, tolerance_c: f32) {
         self.tolerance_c = tolerance_c;
-        self.timer_s = 0.0;
-        self.start_time_s = 0.0;
+        self.timer_s = 0;
+        self.start_time_s = 0;
         self.state = ProfileState::Inactive;
     }
 
-    pub fn start(&mut self, time_s: f32, temperature_c: f32) {
+    pub fn start(&mut self, time_s: u64, temperature_c: f32) {
         if self.state == ProfileState::Inactive {
             if temperature_c < (self.setpoint_c - self.tolerance_c) {
                 self.state = ProfileState::Pending;
@@ -45,7 +45,7 @@ impl TemperatureLevel {
         self.state = ProfileState::Inactive;
     }
 
-    pub fn update(&mut self, time_s: f32, temperature_c: f32) {
+    pub fn update(&mut self, time_s: u64, temperature_c: f32) {
         if self.state == ProfileState::Pending {
             if temperature_c >= (self.setpoint_c - self.tolerance_c) {
                 self.state = ProfileState::Active;
@@ -60,7 +60,7 @@ impl TemperatureLevel {
         }
     }
 
-    pub fn force_active_if_pending(&mut self, time_s: f32) {
+    pub fn force_active_if_pending(&mut self, time_s: u64) {
         if self.state == ProfileState::Pending {
             self.state = ProfileState::Active;
             self.start_time_s = time_s;
@@ -73,10 +73,10 @@ impl TemperatureLevel {
 
 }
 
-#[derive(Clone, Serialize, Deserialize)]
+#[derive(Clone, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct TemperatureProfile {
-    pub start_time_s: f32,
+    pub start_time_s: u64,
     pub tolerance_c: f32,
     pub state: ProfileState,
     pub levels: Vec<TemperatureLevel>,
@@ -86,7 +86,7 @@ pub struct TemperatureProfile {
 impl TemperatureProfile {
     pub fn new() -> TemperatureProfile {
         TemperatureProfile {  
-            start_time_s: 0.0,
+            start_time_s: 0,
             tolerance_c: 0.0,
             state: ProfileState::Inactive,
             levels: Vec::new(),
@@ -94,7 +94,7 @@ impl TemperatureProfile {
         }
     }
 
-    pub fn start(&mut self, time_s: f32, temperature_c: f32) -> f32 {
+    pub fn start(&mut self, time_s: u64, temperature_c: f32) -> f32 {
         if self.state != ProfileState::Inactive {
             if self.active_level >=0 {
                 let level = &self.levels[self.active_level as usize];
@@ -131,7 +131,7 @@ impl TemperatureProfile {
         self.state = ProfileState::Inactive;
     }
 
-    pub fn start_pending_level(&mut self, time_s: f32, temperature_c: f32) {
+    pub fn start_pending_level(&mut self, time_s: u64, temperature_c: f32) {
         if self.active_level >= 0 && self.state != ProfileState::Terminated {
             let level = &mut self.levels[self.active_level as usize];
 
@@ -140,7 +140,7 @@ impl TemperatureProfile {
         }
     }
 
-    pub fn terminate_current_level(&mut self, time_s: f32, temperature_c: f32) {
+    pub fn terminate_current_level(&mut self, time_s: u64, temperature_c: f32) {
         if self.active_level >= 0 && self.state != ProfileState::Terminated {
             let level = &mut self.levels[self.active_level as usize];
 
@@ -150,7 +150,7 @@ impl TemperatureProfile {
     }
 
 
-    pub fn update(&mut self, time_s: f32, temperature_c: f32) -> f32 {
+    pub fn update(&mut self, time_s: u64, temperature_c: f32) -> f32 {
         if self.active_level >= 0 && self.state != ProfileState::Terminated {
 
             let number_of_levels = self.levels.len() as i32;
